// tests/types/InventoryItem.ts
type InventoryItem = {
  Date: string
  Category: string
  Name: string
  Manufacturer: string
  'Model Year': number
  'Part Number': string
  'Bar Code': string
  Description: string
  Location: string
  Taxable: string
  'Unit Cost': string
  'Unit Price': string
  'Last Order': string
  'Reorder Level': number
}

export default InventoryItem